refactor: introduce cli processor package
