+++ cmd/cli/kubectl-kyverno/processor/vap.go
@@ -1,4 +1,4 @@
-package common
+package processor
 
 import (
 	engineapi "github.com/kyverno/kyverno/pkg/engine/api"
@@ -7,10 +7,9 @@ import (
 
 type ValidatingAdmissionPolicies struct{}
 
-func (p *ValidatingAdmissionPolicies) ApplyPolicyOnResource(c ApplyPolicyConfig) ([]engineapi.EngineResponse, error) {
+func (p *ValidatingAdmissionPolicies) ApplyPolicyOnResource(c Processor) ([]engineapi.EngineResponse, error) {
 	engineResp := validatingadmissionpolicy.Validate(c.ValidatingAdmissionPolicy, *c.Resource)
 	ruleResp := engineResp.PolicyResponse.Rules[0]
-
 	if ruleResp.Status() == engineapi.RuleStatusPass {
 		c.Rc.Pass++
 	} else if ruleResp.Status() == engineapi.RuleStatusFail {
@@ -18,6 +17,5 @@ func (p *ValidatingAdmissionPolicies) ApplyPolicyOnResource(c ApplyPolicyConfig)
 	} else if ruleResp.Status() == engineapi.RuleStatusError {
 		c.Rc.Error++
 	}
-
 	return []engineapi.EngineResponse{engineResp}, nil
 }
