--- cmd/cli/kubectl-kyverno/utils/common/policy_interface.go
@@ -1,9 +1,0 @@
-package common
-
-import (
-	engineapi "github.com/kyverno/kyverno/pkg/engine/api"
-)
-
-type PolicyInterface interface {
-	ApplyPolicyOnResource(c ApplyPolicyConfig) ([]engineapi.EngineResponse, error)
-}
